fix(front): simplify signup flow
